+++ src/components/ListOfProjectsItems/index.tsx
