+++ src/components/ListOfProjectsItems/style.tsx
