+++ src/containers/content.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { ContainerWrap } from './style';
-import { List } from '../components/List';
+import { List } from '../components/ListOfProjects';
 import { HeaderNav }  from '../components/Header';
